major refactor 1/2 way done-- using proxy objects to access the database
