+++ app/data/__init__.py
@@ -3,11 +3,9 @@
 The data module contains exactly what you'd expect: everything related to data
 processing, collection, and storage.
 """
+from .globals import cache
 from .database import db
-from .database import cache
 from .celery import celery_app
-
-db.metadata.clear()
 
 # SqlAlchemy database models
 from .models.boathouse import Boathouse
