+++ app/data/globals.py
@@ -3,32 +3,45 @@ from typing import List
 from werkzeug.local import LocalProxy
 from flask import g
 
+from flask_caching import Cache as _Cache
 from app.data.models.website_options import WebsiteOptions
 from app.data.models.boathouse import Boathouse
 from app.data.models.reach import Reach
-from app.data.database import cache
 
 
-def request_context_object(func: callable, key: str) -> Any:
-    """Factory for implementing a multilayer cache for a global object accessed
-    via the Postgres database.
-
-    The object is cached both in memory during the request context, and in the
-    Redis layer. So when the request dies, the object still persists in Redis,
-    and will be reloaded into a new request context.
+class Cache(_Cache):
+    """Implementation of the cache that also handles the cached_proxy objects.
     """
+
+    app_context_variables: List[str]
+
+    def __init__(self):
+        super().__init__()
+        self.app_context_variables = []
+
+    def clear(self) -> None:
+        for i in self.app_context_variables:
+            if i in g:
+                g.pop(i)
+        super().clear()
+
+
+cache = Cache()
+
+
+def cached_proxy(func: callable, key: str) -> Any:
+    """Factory for implementing a cache for a global object accessed via the
+    Postgres database.
+    """
+
+    cache.app_context_variables.append(key)
 
     def _fetch():
         active = g.get(key)
         if active:
             return active
-        elif cache.cache.has(key):
-            res = cache.get(key)
-            g.setdefault(key, res)
-            return res
         else:
             res = func()
-            cache.set(key, res)
             g.setdefault(key, res)
             return res
 
@@ -36,10 +49,10 @@ def request_context_object(func: callable, key: str) -> Any:
 
 
 website_options: WebsiteOptions = \
-    request_context_object(WebsiteOptions.get, key='website_options')  # type: ignore
+    cached_proxy(WebsiteOptions.get, key='website_options')  # type: ignore
 
 boathouses: List[Boathouse] = \
-    request_context_object(Boathouse.get_all, key='boathouse_list')  # type: ignore
+    cached_proxy(Boathouse.get_all, key='boathouse_list')  # type: ignore
 
 reaches: List[Reach] = \
-    request_context_object(Reach.get_all, key='reach_list')  # type: ignore
+    cached_proxy(Reach.get_all, key='reach_list')  # type: ignore
